refactor: move macOS parent-death watchdog into MLX backend
Encapsulate the kqueue-based watchdog in
hardware_backend/mlx/parent_watchdog.py and have the darwin branch of
kill_itself_when_parent_died() just call start_parent_death_watcher(),
keeping the platform dispatch in common.py clean.
